size_t/ptrdiff_t transition part 5: the libast zakkaya (#954)
This is the fifth of the thickfold patch series, which enables
ksh93 to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The libast compatibility functions.
- The libast miscellaneous functions (e.g. optget and friends).
- The libast portability functions.

Remarks:
- The second argument of getfsstat can either be a long (FreeBSD)
  or a size_t (OpenBSD). This patch opts to cast the second
  argument as a size_t, which will produce a warning on FreeBSD and
  no warning on OpenBSD.
- Fixed a misplaced noreturn attribute for _ast_assertfail and
  added noreturn to debug_fatal() (whichs ends with abort()).

- Marked error_break(), typefix(), and various CDT functions as
  static to reflect actual usage.

Change in the number of warnings on Linux when compiling with clang
using -Wsign-compare -Wshorten-64-to-32 -Wsign-conversion
-Wimplicit-int-conversion: 2,450 => 1,946 => 37 (progression from
part 4 => part 5 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/comp/frexpl.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 
@@ -47,7 +48,7 @@ static long double	pow2tab[LDBL_MAX_EXP + 1];
 static int
 init(void)
 {
-	int		x;
+	size_t		x;
 	long double	g;
 
 	g = 1;
